+++ src/components/index.ts
@@ -1,0 +1,9 @@
+import Navbar from "./Navbar";
+import Container from "./Container";
+
+const exportedComponents = {
+  Navbar,
+  Container,
+};
+
+export default exportedComponents;